CLI enhancements for both gamefic and gfk commands
* Moved from using slop to thor, since that was easier and provided an easy DSL for commands with
options and builtin help pages for all of these.
* Removed some unnecessary code from lib/gamefic/shell and lib/gamefic-sdk/shell and refactored the remains,
as the new thor commands are simple enough to be a few lines of code.
* Fixed two of the specs (one got only style fixes, other got one test for some old method removed,
since the method was removed as well).
* Verified locally that all the commands are still usable and working as before.

I've also made two changes that are a bit more controversial, but not really very dramatic:

* Updated the required Ruby version to 2.1. The reasoning is that it allows to use the keyword arguments and
it's a good thing as it allows to do `def foo(bar:, baz: false)` so you can call `foo baz: true, bar: 1` or
`foo bar: 1` and all of this will be valid.
* Added the `.rubocop.yml` config file with a few exceptions, so my editor will not scream at me so much.